requirements.txt
pyTelegramBotAPI
pyowm
psycopg2-binary
flask
requests
waitress